Added script to process Hanno Böck's AXFR scans into SQLite
This will serve the basis to generate new wordlists in a reproduceable
and more scientific manner. For example, generating the final result
will be possible with:

 SELECT nom, GROUP_CONCAT(rec), SUM(cnt) as total_cnt FROM axfr_counts GROUP BY nom ORDER BY total_cnt DESC LIMIT 10;

This list should be used to supplement the existing included lists,
it will provide hints for the types of records which exist, e.g.

	_ldap._tcp.default-first-site-name|SRV|38
	admin|A,AAAA,CNAME,HINFO,MX,NS,TXT|1259
	mail|A,AAAA,CNAME,MX,NS,NSEC,PTR,RRSIG,SPF,TXT|23268
	_msdcs|A,AAAA,CNAME,NS,SRV|161

The record type hints can be used to perform additional queries
other than the most common ones (e.g. A, AAAA and CNAME) without
increasing the total number of queries N-fold.